Split out the TimeSource for generic use in other tests
